+++ cmn-data-svc-server/src/main/java/org/acumos/cds/service/AbstractSearchServiceImpl.java
@@ -38,26 +38,24 @@ public abstract class AbstractSearchServiceImpl {
 	 * @param criteria
 	 *            Criteria object to extend
 	 * @param queryParameters
-	 *            Map of field name - field value pairs; ignored if null or empty.
-	 *            An "equals" criterion is created, unless the field value is n
-	 *            array, in which case a "in" criterion is created.
+	 *            Map of field name - field value pairs. Value may be a scalar or
+	 *            array. An "equals" criterion is created for scalar values. An "in"
+	 *            criterion is created for array values.
 	 * @param isOr
 	 *            If true, treat the query as a disjunction; else as a conjunction.
 	 */
 	protected void buildCriteria(Criteria criteria, Map<String, ? extends Object> queryParameters, boolean isOr) {
-		if (queryParameters != null && queryParameters.size() > 0) {
-			Junction junction = isOr ? Restrictions.disjunction() : Restrictions.conjunction();
-			criteria.add(junction);
-			for (Map.Entry<String, ? extends Object> entry : queryParameters.entrySet()) {
-				Criterion criterion = null;
-				if (entry.getValue().getClass().isArray()) {
-					Object[] array = (Object[]) entry.getValue();
-					criterion = Restrictions.in(entry.getKey(), array);
-				} else {
-					criterion = Restrictions.eq(entry.getKey(), entry.getValue());
-				}
-				junction.add(criterion);
+		Junction junction = isOr ? Restrictions.disjunction() : Restrictions.conjunction();
+		criteria.add(junction);
+		for (Map.Entry<String, ? extends Object> entry : queryParameters.entrySet()) {
+			Criterion criterion = null;
+			if (entry.getValue().getClass().isArray()) {
+				Object[] array = (Object[]) entry.getValue();
+				criterion = Restrictions.in(entry.getKey(), array);
+			} else {
+				criterion = Restrictions.eq(entry.getKey(), entry.getValue());
 			}
+			junction.add(criterion);
 		}
 	}
 }