+++ cmn-data-svc-server/src/main/java/org/acumos/cds/service/UserSearchService.java
@@ -31,16 +31,16 @@ import org.acumos.cds.domain.MLPUser;
 public interface UserSearchService {
 
 	/**
-	 * Gets all instances matching all query parameters; i.e., the conditions are
-	 * AND-ed together.
+	 * Searches for instances matching all or one of the query parameters, depending
+	 * on the isOr parameter.
 	 * 
 	 * @param queryParameters
-	 *            field-name, field-value pairs; ignored if null or empty.
+	 *            field-name, field-value pairs.  Value may be scalar or array.
 	 * @param isOr
 	 *            If true, the query is a disjunction ("or"); otherwise the query is
 	 *            a conjunction ("and").
 	 * @return List of instances, which may be empty.
 	 */
-	List<MLPUser> getUsers(Map<String, ? extends Object> queryParameters, boolean isOr);
+	List<MLPUser> findUsers(Map<String, ? extends Object> queryParameters, boolean isOr);
 
 }